+++ lib/src/widgets/error_widget.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class SomethingWentWrong extends StatelessWidget {
+  const SomethingWentWrong({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Padding(
+        padding: EdgeInsets.all(12.0),
+        child: Text('Something went wrong!'),
+      ),
+    );
+  }
+}